+++ user-apps/FlexiSheet/Application/Inspection/FSTableInspectorPane.m
@@ -5,7 +5,7 @@
 //  Created by Stefan Leuker on 15-DEC-2001.
 //  Copyright (c) 2001-2003 Stefan Leuker. All rights reserved.
 //
-//  $Id: FSTableInspectorPane.m,v 1.1 2008/10/28 13:10:29 hns Exp $
+//  $Id: FSTableInspectorPane.m,v 1.2 2014/01/26 09:23:52 buzzdee Exp $
 
 #import "FSTableInspectorPane.h"
 
@@ -89,7 +89,7 @@
 // TableDataSource
 //
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     if (aTableView == categories) {
         return [[table headers] count];
@@ -98,7 +98,9 @@
 }
 
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)column row:(int)row
+- (id)tableView:(NSTableView *)aTableView
+	objectValueForTableColumn:(NSTableColumn *)column
+			      row:(NSInteger)row
 {
     if (aTableView == categories) {
         return [[[table headers] objectAtIndex:row] label];
